pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bird-vision"
version = "0.1.0"
description = "Multi-modal bird identification with computer vision and audio analysis"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "deeplake>=3.8.0",
    "timm>=0.9.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.8.0",
    "wandb>=0.16.0",
    "albumentations>=1.3.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "tensorboard>=2.14.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "openvino>=2023.3.0",
    "coremltools>=7.0",
    "tensorflow-lite>=2.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]
mobile = [
    "torch-mobile>=0.1.0",
    "torchscript>=1.0.0",
]
audio = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "audioread>=3.0.0",
]
esp32 = [
    "thop>=0.1.1",  # For FLOPS calculation
    "platformio>=6.0.0",  # For ESP32 development
]

[project.scripts]
bird-vision = "bird_vision.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=bird_vision --cov-report=term-missing"
testpaths = ["tests"]